requirements.txt
PyPDF4
pdfminer.six
pandas
python-dotenv
streamlit
